shamefullyAdjustGeo
