+++ src/components/index.js
@@ -53,6 +53,7 @@ import {
   AxesFold,
   Fold,
   LayoutPanel,
+  LayoutSection,
   Panel,
   Section,
   TraceAccordion,
@@ -87,6 +88,7 @@ export {
   ArrowSelector,
   AxesFold,
   AxesRange,
+  LayoutSection,
   AxesSelector,
   Button,
   CanvasSize,
