+++ src/index.js
@@ -51,6 +51,7 @@ import {
   PositioningNumeric,
   NumericFractionInverse,
   LayoutPanel,
+  LayoutSection,
   LineDashSelector,
   LineShapeSelector,
   Numeric,
@@ -121,6 +122,7 @@ export {
   PositioningNumeric,
   NumericFractionInverse,
   LayoutPanel,
+  LayoutSection,
   LineDashSelector,
   LineShapeSelector,
   Numeric,
